commit on in progress restructure of how settlements work across contracts, using a required position-modification triggered delta that can be modified by MM actions

+++ solidity/src/types/VTS.sol
@@ -7,8 +7,8 @@ import {PositionId} from "./Position.sol";
 struct TokenConfiguration {
     // Grace period time
     uint256 gracePeriodTime;
-    // Base VTS Rate
-    uint256 baseVTSRate;
+    // Base VTS Rate in bps (basis points)
+    uint16 baseVTSRate; // 1000 = 10% | 10000 = 100%
 }
 
 struct MarketVTSConfiguration {
